Add LBP implementation.
I have copied and pasted everything... and it now works?
I really don't know what the difference between this branch
and local_binary_pattern is, and the git diff looks to be all whitespace
differences.

I'm really confused, but for times sake I'm not going to poke the bear.
I am going to flush out the tests and verify the correctness of the algo.